Exp4/generation_s2/Stegano/_m3_requirements_raw.txt
<requirements>
Pillow
piexif
</requirements>